refactor: LottoNumber 생성자의 파라미터로 LottoNumberGeneratorStrategy받도록 수정, LottoNumberTest 클래스 분리
